+++ playground/Lambda/Lambda.AppHost/Program.cs
@@ -1,4 +1,5 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+using Amazon.CDK.AWS.DynamoDB;
 using Amazon.Lambda;
 using Aspire.Hosting.AWS.Lambda;
 
@@ -11,6 +12,13 @@ cdkStackResource.WithTag("aws-repo", "integrations-on-dotnet-aspire-for-aws");
 
 var sqsDemoQueue1 = cdkStackResource.AddSQSQueue("DemoQueue1");
 var sqsDemoQueue2 = cdkStackResource.AddSQSQueue("DemoQueue2");
+
+var table = cdkStackResource.AddDynamoDBTable("DemoTable", new TableProps
+{
+    PartitionKey = new Amazon.CDK.AWS.DynamoDB.Attribute { Name = "Id", Type = AttributeType.STRING },
+    Stream = StreamViewType.NEW_AND_OLD_IMAGES,
+    BillingMode = BillingMode.PAY_PER_REQUEST
+});
 
 builder.AddAWSLambdaFunction<Projects.ToUpperLambdaFunctionExecutable>("ToUpperFunction", lambdaHandler: "ToUpperLambdaFunctionExecutable", new LambdaFunctionOptions { ApplicationLogLevel = ApplicationLogLevel.DEBUG, LogFormat = LogFormat.JSON });
 
@@ -41,6 +49,13 @@ builder.AddAWSLambdaFunction<Projects.SQSProcessorFunction>("SQSProcessorFunctio
         // CDK output parameters are not attempted to be added.
         .WithReference(sqsDemoQueue1)
         .WithReference(sqsDemoQueue2);
+
+
+builder.AddAWSLambdaFunction<Projects.DynamoDBProcessorFunction>("DynamoDBProcessorFunction", "DynamoDBProcessorFunction::DynamoDBProcessorFunction.Function::FunctionHandler")
+        .WithDynamoDBStreamsEventSource(table, new DynamoDBStreamsEventSourceOptions
+        {
+            BatchSize = 5
+        });
 
 
 builder.Build().Run();